pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "synth-managed-research"
version = "0.1.0"
description = "Public SDK, MCP, and schemas for Managed Research."
readme = "PACKAGE_README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Synth Laboratories" }]
dependencies = [
    "httpx>=0.28.1",
    "pynacl>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "build>=1.2.2.post1",
    "pytest>=8.3.3",
    "ruff>=0.15.0",
    "ty>=0.0.28",
]

[project.urls]
Homepage = "https://github.com/synth-laboratories/managed-research"
Repository = "https://github.com/synth-laboratories/managed-research"
Issues = "https://github.com/synth-laboratories/managed-research/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["managed_research", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.hatch.build.targets.wheel]
packages = ["managed_research"]

[dependency-groups]
dev = [
    "build>=1.2.2.post1",
    "pytest>=8.3.3",
    "ruff>=0.15.0",
    "ty>=0.0.28",
]
